Fix the `keys` method for PollSelector
The current implementation of the `keys` method is broken, because it can return
null values in the resulted set. For instance, `PollSelector` starts with `keyArray`
of  the size of 1. The returned set then will always have a null value in the set.
External systems that call this method don't expect that the set will contain null
values and don't check for it. For instance, Jetty doesn't do that.