Always clear out descriptor metadata on registration
When registering on a channel, we should always reset the descriptor to its initial state. It is quite possible for a file descriptor to be re-used. In such scenarios, without this change, events may not fire. I observed this situation exactly when using jnr-unixsocket. I cancelled a key for OP_ACCEPT and closed the channel. On quickly opening a new channel and registering for OP_ACCEPT, the OS returned the same file descriptor as before.

The changes here have been tested locally on my code based and observed to work correctly.